+++ src/officialeye/_api/template/supervisor.py
@@ -1,7 +1,16 @@
-from abc import ABC, abstractmethod
+from __future__ import annotations
 
+from abc import ABC, abstractmethod
+from typing import Iterable, TYPE_CHECKING
+
+from officialeye._api.template.matching_result import IMatchingResult
+from officialeye._api.template.supervision_result import SupervisionResult
 from officialeye._api.template.template_interface import ITemplate
 from officialeye._api.config import SupervisorConfig
+
+
+if TYPE_CHECKING:
+    from officialeye.types import ConfigDict
 
 
 class ISupervisor(ABC):
@@ -12,5 +21,22 @@ class ISupervisor(ABC):
         raise NotImplementedError()
 
     @abstractmethod
-    def setup(self, template: ITemplate, matching_result, /) -> None:
+    def setup(self, template: ITemplate, matching_result: IMatchingResult, /) -> None:
         raise NotImplementedError()
+
+    @abstractmethod
+    def supervise(self, template: ITemplate, matching_result: IMatchingResult, /) -> Iterable[SupervisionResult]:
+        raise NotImplementedError()
+
+
+class Supervisor(ISupervisor, ABC):
+
+    def __init__(self, supervisor_id: str, config_dict: ConfigDict, /):
+        super().__init__()
+
+        self._supervisor_id = supervisor_id
+        self._config = SupervisorConfig(config_dict, supervisor_id)
+
+    @property
+    def config(self) -> SupervisorConfig:
+        return self._config