+++ src/officialeye/_api_builtins/init.py
@@ -12,11 +12,15 @@ from officialeye._api_builtins.mutator.clahe import CLAHEMutator
 from officialeye._api_builtins.mutator.grayscale import GrayscaleMutator
 from officialeye._api_builtins.mutator.non_local_means_denoising import NonLocalMeansDenoisingMutator
 from officialeye._api_builtins.mutator.rotate import RotateMutator
+from officialeye._api_builtins.supervisor.combinatorial import CombinatorialSupervisor
+from officialeye._api_builtins.supervisor.least_squares_regression import LeastSquaresRegressionSupervisor
 
 if TYPE_CHECKING:
     from officialeye.types import ConfigDict
     # noinspection PyProtectedMember
     from officialeye._api.context import Context
+    # noinspection PyProtectedMember
+    from officialeye._api.template.supervisor import ISupervisor
 
 
 # mutator generators
@@ -43,6 +47,15 @@ def _gen_matcher_sift_flann(config: ConfigDict, /) -> IMatcher:
     return SiftFlannMatcher(config)
 
 
+# supervisor generators
+def _gen_supervisor_combinatorial(config: ConfigDict, /) -> ISupervisor:
+    return CombinatorialSupervisor(config)
+
+
+def _gen_supervisor_least_squares_regression(config: ConfigDict, /) -> ISupervisor:
+    return LeastSquaresRegressionSupervisor(config)
+
+
 def initialize_builtins(context: Context, /):
 
     # register mutators
@@ -53,3 +66,7 @@ def initialize_builtins(context: Context, /):
 
     # register matchers
     context.register_matcher(SiftFlannMatcher.MATCHER_ID, _gen_matcher_sift_flann)
+
+    # register supervisors
+    context.register_supervisor(CombinatorialSupervisor.SUPERVISOR_ID, _gen_supervisor_combinatorial)
+    context.register_supervisor(LeastSquaresRegressionSupervisor.SUPERVISOR_ID, _gen_supervisor_combinatorial)