+++ src/officialeye/_api_builtins/supervisor/combinatorial.py
@@ -1,23 +1,38 @@
+from __future__ import annotations
+
 import random
-from typing import Dict, Generator
+from typing import Dict, TYPE_CHECKING, Iterable, List
 
 import numpy as np
 import z3
 
+# noinspection PyProtectedMember
+from officialeye._api.template.template_interface import ITemplate
+# noinspection PyProtectedMember
+from officialeye._api.template.match import IMatch
+# noinspection PyProtectedMember
+from officialeye._api.template.matching_result import IMatchingResult
+# noinspection PyProtectedMember
+from officialeye._api.template.supervisor import Supervisor
+# noinspection PyProtectedMember
+from officialeye._api.template.supervision_result import SupervisionResult
+# noinspection PyProtectedMember
+from officialeye._internal.context.singleton import get_internal_afi
+# noinspection PyProtectedMember
+from officialeye._internal.feedback.verbosity import Verbosity
 from officialeye.error.errors.supervision import ErrSupervisionInvalidEngineConfig
 
-from officialeye._internal.matching.match import Match
-from officialeye._internal.template.matcher_result import InternalMatcherResult
-from officialeye._internal.supervision.result import SupervisionResult
-from officialeye._internal.supervision.supervisor import Supervisor
+
+if TYPE_CHECKING:
+    from officialeye.types import ConfigDict
 
 
 class CombinatorialSupervisor(Supervisor):
 
-    ENGINE_ID = "combinatorial"
+    SUPERVISOR_ID = "combinatorial"
 
-    def __init__(self, template_id: str, kmr: InternalMatcherResult, /):
-        super().__init__(CombinatorialSupervisor.ENGINE_ID, template_id, kmr)
+    def __init__(self, config_dict: ConfigDict, /):
+        super().__init__(CombinatorialSupervisor.SUPERVISOR_ID, config_dict)
 
         # setup configuration
         def _min_match_factor_preprocessor(v: str) -> float:
@@ -26,19 +41,19 @@ class CombinatorialSupervisor(Supervisor):
 
             if v > 1.0:
                 raise ErrSupervisionInvalidEngineConfig(
-                    f"while loading the '{CombinatorialSupervisor.ENGINE_ID}' supervisor",
+                    f"while loading the '{CombinatorialSupervisor.SUPERVISOR_ID}' supervisor",
                     f"The `min_match_factor` value ({v}) cannot exceed 1.0."
                 )
 
             if v < 0.0:
                 raise ErrSupervisionInvalidEngineConfig(
-                    f"while loading the '{CombinatorialSupervisor.ENGINE_ID}' supervisor",
+                    f"while loading the '{CombinatorialSupervisor.SUPERVISOR_ID}' supervisor",
                     f"The `min_match_factor` value ({v}) cannot be negative."
                 )
 
             return v
 
-        self.get_config().set_value_preprocessor("min_match_factor", _min_match_factor_preprocessor)
+        self.config.set_value_preprocessor("min_match_factor", _min_match_factor_preprocessor)
 
         def _max_transformation_error_preprocessor(v: str) -> int:
 
@@ -46,19 +61,19 @@ class CombinatorialSupervisor(Supervisor):
 
             if v < 0:
                 raise ErrSupervisionInvalidEngineConfig(
-                    f"while loading the '{CombinatorialSupervisor.ENGINE_ID}' supervisor.",
+                    f"while loading the '{CombinatorialSupervisor.SUPERVISOR_ID}' supervisor.",
                     f"The `max_transformation_error` value ({v}) cannot be negative."
                 )
 
             if v > 5000:
                 raise ErrSupervisionInvalidEngineConfig(
-                    f"while loading the '{CombinatorialSupervisor.ENGINE_ID}' supervisor.",
+                    f"while loading the '{CombinatorialSupervisor.SUPERVISOR_ID}' supervisor.",
                     f"The `max_transformation_error` value ({v}) is too high."
                 )
 
             return v
 
-        self.get_config().set_value_preprocessor("max_transformation_error", _max_transformation_error_preprocessor)
+        self.config.set_value_preprocessor("max_transformation_error", _max_transformation_error_preprocessor)
 
         def _z3_timeout_preprocessor(v: str) -> int:
 
@@ -66,38 +81,49 @@ class CombinatorialSupervisor(Supervisor):
 
             if v < 1:
                 raise ErrSupervisionInvalidEngineConfig(
-                    f"while loading the '{CombinatorialSupervisor.ENGINE_ID}' supervisor.",
+                    f"while loading the '{CombinatorialSupervisor.SUPERVISOR_ID}' supervisor.",
                     f"The `z3_timeout` value ({v}) cannot be negative or zero."
                 )
 
             return v
 
-        self.get_config().set_value_preprocessor("z3_timeout", _z3_timeout_preprocessor)
+        self.config.set_value_preprocessor("z3_timeout", _z3_timeout_preprocessor)
 
         # initialize all engine-specific values
-        self._z3_context = z3.Context()
+        self._z3_context: z3.Context | None = None
 
         # create variables for components of the translation matrix
+        self._transformation_matrix: np.ndarray | None = None
+
+        # keys: matches (instances of Match)
+        # values: z3 integer variables representing the errors for each match,
+        # i.e., how consistent the match is with the affine transformation model
+        self._match_weight: Dict[IMatch, z3.ArithRef] = {}
+
+        self._minimum_weight_to_enforce: float | None = None
+
+        self._max_transformation_error = self.config.get("max_transformation_error")
+
+    def setup(self, template: ITemplate, matching_result: IMatchingResult, /) -> None:
+        self._z3_context = z3.Context()
+
         self._transformation_matrix = np.array([
             [z3.Real("a", ctx=self._z3_context), z3.Real("b", ctx=self._z3_context)],
             [z3.Real("c", ctx=self._z3_context), z3.Real("d", ctx=self._z3_context)]
         ], dtype=z3.AstRef)
 
-        # keys: matches (instances of Match)
-        # values: z3 integer variables representing the errors for each match,
-        # i.e., how consistent the match is with the affine transformation model
-        self._match_weight: Dict[Match, z3.ArithRef] = {}
+        for match in matching_result.get_all_matches():
+            self._match_weight[match] = z3.Real(
+                f"w_{match.template_point[0]}_{match.template_point[1]}_{match.target_point[0]}_{match.target_point[1]}",
+                ctx=self._z3_context
+            )
 
-        for match in self._kmr.get_matches():
-            self._match_weight[match] = z3.Real(f"w_{match.get_debug_identifier()}", ctx=self._z3_context)
+        # calculate the minimum weight that we need to enforce
+        _config_min_match_factor = self.config.get("min_match_factor", default=0.1)
 
-        _config_min_match_factor = self.get_config().get("min_match_factor", default=0.1)
-        get_logger().debug(f"Min match factor: {_config_min_match_factor}")
+        self._minimum_weight_to_enforce = matching_result.get_total_match_count() * _config_min_match_factor
 
-        self._minimum_weight_to_enforce = self._kmr.get_total_match_count() * _config_min_match_factor
-        self._max_transformation_error = self.get_config().get("max_transformation_error")
-
-    def _get_consistency_check(self, match: Match, delta: np.ndarray, delta_prime: np.ndarray, /) -> z3.AstRef:
+    def _get_consistency_check(self, match: IMatch, delta: np.ndarray, delta_prime: np.ndarray, /) -> z3.AstRef:
         """
         Generates a z3 formula asserting the consistency of the match with the affine linear transformation model.
         Consistency does not mean ideal matching of coordinates; rather, the template position with the affine
@@ -114,7 +140,7 @@ class CombinatorialSupervisor(Supervisor):
         translated_template_point = self._transformation_matrix @ (template_point - delta) + delta_prime
         translated_template_point_x, translated_template_point_y = translated_template_point
 
-        target_point_x, target_point_y = match.get_target_point()
+        target_point_x, target_point_y = match.target_point
 
         return z3.And(
             translated_template_point_x - target_point_x <= self._max_transformation_error,
@@ -123,15 +149,15 @@ class CombinatorialSupervisor(Supervisor):
             target_point_y - translated_template_point_y <= self._max_transformation_error,
         )
 
-    def _run(self) -> Generator[SupervisionResult, None, None]:
+    def supervise(self, template: ITemplate, matching_result: IMatchingResult, /) -> Iterable[SupervisionResult]:
 
-        weights_lower_bounds = z3.And(*(self._match_weight[match] >= 0 for match in self._kmr.get_matches()), self._z3_context)
-        weights_upper_bounds = z3.And(*(self._match_weight[match] <= 1 for match in self._kmr.get_matches()), self._z3_context)
+        weights_lower_bounds = z3.And(*(self._match_weight[match] >= 0 for match in matching_result.get_all_matches()), self._z3_context)
+        weights_upper_bounds = z3.And(*(self._match_weight[match] <= 1 for match in matching_result.get_all_matches()), self._z3_context)
 
-        total_weight = z3.Sum(*(self._match_weight[match] for match in self._kmr.get_matches()))
+        total_weight = z3.Sum(*(self._match_weight[match] for match in matching_result.get_all_matches()))
 
         solver = z3.Optimize(ctx=self._z3_context)
-        solver.set("timeout", self.get_config().get("z3_timeout", default=2500))
+        solver.set("timeout", self.config.get("z3_timeout", default=2500))
 
         solver.add(weights_lower_bounds)
         solver.add(weights_upper_bounds)
@@ -139,21 +165,21 @@ class CombinatorialSupervisor(Supervisor):
 
         solver.maximize(total_weight)
 
-        for keypoint_id in self._kmr.get_keypoint_ids():
-            keypoint_matches = list(self._kmr.matches_for_keypoint(keypoint_id))
+        for keypoint in template.keypoints:
+            keypoint_matches: List[IMatch] = list(matching_result.get_matches_for_keypoint(keypoint.identifier))
 
             if len(keypoint_matches) == 0:
                 continue
 
             # TODO: think whether this is a good algorithm design decision, and improve it if not
-            anchor_match = keypoint_matches[random.randint(0, len(keypoint_matches) - 1)]
+            anchor_match: IMatch = random.choice(keypoint_matches)
 
             delta = anchor_match.get_original_template_point()
-            delta_prime = anchor_match.get_target_point()
+            delta_prime = anchor_match.target_point
 
             solver.push()
 
-            for match in self._kmr.get_matches():
+            for match in matching_result.get_all_matches():
                 solver.add(z3.Implies(
                     self._match_weight[match] > 0,
                     # consistency check
@@ -164,12 +190,12 @@ class CombinatorialSupervisor(Supervisor):
             result = solver.check()
 
             if result == z3.unsat:
-                get_logger().warn("Could not satisfy the imposed constraints.", fg="red")
+                get_internal_afi().warn(Verbosity.INFO_VERBOSE, "Could not satisfy the imposed constraints.")
                 solver.pop()
                 continue
 
             if result == z3.unknown:
-                get_logger().warn("Could not decide the satifiability of the imposed constraints.", fg="red")
+                get_internal_afi().warn(Verbosity.INFO, "Could not decide the satifiability of the imposed constraints.")
                 solver.pop()
                 continue
 
@@ -184,15 +210,22 @@ class CombinatorialSupervisor(Supervisor):
             # extract transformation matrix from model
             transformation_matrix = model_evaluator(self._transformation_matrix)
 
-            _result = SupervisionResult(self.template_id, self._kmr, delta, delta_prime, transformation_matrix)
-            # add fixed constant to make sure that the score value is always non-negative
-            _result.set_score(model_total_weight)
+            _result = SupervisionResult(
+                matching_result,
+                delta=delta,
+                delta_prime=delta_prime,
+                transformation_matrix=transformation_matrix,
+                score=model_total_weight
+            )
 
-            for match in self._kmr.get_matches():
+            for match in matching_result.get_all_matches():
                 match_weight = model_evaluator(self._match_weight[match])
                 _result.set_match_weight(match, match_weight)
 
-            get_logger().debug(f"Error: {_result.get_weighted_mse()} Total weight and score: {model_total_weight}")
+            get_internal_afi().info(
+                Verbosity.INFO_VERBOSE,
+                f"Error: {_result.get_weighted_mse()} Total weight and score: {model_total_weight}"
+            )
 
             yield _result