+++ src/officialeye/error/errors/general.py
@@ -34,3 +34,25 @@ class ErrInvalidKey(ErrGeneral):
 
     def __reduce__(self):
         return self.__class__, self._init_args
+
+
+class ErrInvalidIdentifier(ErrGeneral):
+
+    def __init__(self, while_text: str, problem_text: str, /):
+        super().__init__(while_text, problem_text)
+
+        self._init_args = while_text, problem_text
+
+    def __reduce__(self):
+        return self.__class__, self._init_args
+
+
+class ErrObjectNotInitialized(ErrGeneral):
+
+    def __init__(self, while_text: str, problem_text: str, /):
+        super().__init__(while_text, problem_text)
+
+        self._init_args = while_text, problem_text
+
+    def __reduce__(self):
+        return self.__class__, self._init_args